+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This file is used to list changes made in each version of the devpi cookbook.
 - Fully embrace ChefDK for development.
 - Switch attribute access from symbols to strings.
 - Remove runit and supervisor recipes.
-- Add `devpi_server` LWRP.
+- Add `devpi_server` and `devpi_nginx_site` LWRPs.
 
 ## 1.1.0 (Never Released)
 - Add the *client* recipe that installs the `devpi` command line utility.